requirements.txt
ssdeep
python-magic
Flask
Flask-SqlAlchemy
python-dateutil
yara-python
nose2
psycopg2
